`shannon_entropy` from `utils` to `entropies`
