+++ src/qibo/quantum_info/utils.py
@@ -127,72 +127,6 @@ def hadamard_transform(array, implementation: str = "fast", backend=None):
     return array
 
 
-def shannon_entropy(probability_array, base: float = 2, backend=None):
-    """Calculate the Shannon entropy of a probability array :math:`\\mathbf{p}`, which is given by
-
-    .. math::
-        H(\\mathbf{p}) = - \\sum_{k = 0}^{d^{2} - 1} \\, p_{k} \\, \\log_{b}(p_{k}) \\, ,
-
-    where :math:`d = \\text{dim}(\\mathcal{H})` is the dimension of the
-    Hilbert space :math:`\\mathcal{H}`, :math:`b` is the log base (default 2),
-    and :math:`0 \\log_{b}(0) \\equiv 0`.
-
-    Args:
-        probability_array (ndarray or list): a probability array :math:`\\mathbf{p}`.
-        base (float): the base of the log. Defaults to  :math:`2`.
-        backend (:class:`qibo.backends.abstract.Backend`, optional): backend to be used
-            in the execution. If ``None``, it uses :class:`qibo.backends.GlobalBackend`.
-            Defaults to ``None``.
-
-    Returns:
-        (float): The Shannon entropy :math:`H(\\mathcal{p})`.
-    """
-    if backend is None:  # pragma: no cover
-        backend = GlobalBackend()
-
-    if isinstance(probability_array, list):
-        probability_array = backend.cast(probability_array, dtype=float)
-
-    if base <= 0:
-        raise_error(ValueError, "log base must be non-negative.")
-
-    if len(probability_array.shape) != 1:
-        raise_error(
-            TypeError,
-            f"Probability array must have dims (k,) but it has {probability_array.shape}.",
-        )
-
-    if len(probability_array) == 0:
-        raise_error(TypeError, "Empty array.")
-
-    if any(probability_array < 0) or any(probability_array > 1.0):
-        raise_error(
-            ValueError,
-            "All elements of the probability array must be between 0. and 1..",
-        )
-
-    if np.abs(np.sum(probability_array) - 1.0) > PRECISION_TOL:
-        raise_error(ValueError, "Probability array must sum to 1.")
-
-    if base == 2:
-        log_prob = np.where(probability_array != 0.0, np.log2(probability_array), 0.0)
-    elif base == 10:
-        log_prob = np.where(probability_array != 0, np.log10(probability_array), 0.0)
-    elif base == np.e:
-        log_prob = np.where(probability_array != 0, np.log(probability_array), 0.0)
-    else:
-        log_prob = np.where(
-            probability_array != 0, np.log(probability_array) / np.log(base), 0.0
-        )
-
-    entropy = -np.sum(probability_array * log_prob)
-
-    # absolute value if entropy == 0.0 to avoid returning -0.0
-    entropy = np.abs(entropy) if entropy == 0.0 else entropy
-
-    return complex(entropy).real
-
-
 def hellinger_distance(prob_dist_p, prob_dist_q, validate: bool = False, backend=None):
     """Calculate the Hellinger distance :math:`H(p, q)` between
     two discrete probability distributions, :math:`\\mathbf{p}` and :math:`\\mathbf{q}`.
