+++ tests/test_quantum_info_entropies.py
@@ -2,8 +2,48 @@ import numpy as np
 import pytest
 
 from qibo.config import PRECISION_TOL
-from qibo.quantum_info.entropies import entanglement_entropy, entropy
+from qibo.quantum_info.entropies import entanglement_entropy, entropy, shannon_entropy
 from qibo.quantum_info.random_ensembles import random_statevector, random_unitary
+
+
+def test_shannon_entropy_errors(backend):
+    with pytest.raises(ValueError):
+        prob = np.array([1.0, 0.0])
+        prob = backend.cast(prob, dtype=prob.dtype)
+        test = shannon_entropy(prob, -2, backend=backend)
+    with pytest.raises(TypeError):
+        prob = np.array([[1.0], [0.0]])
+        prob = backend.cast(prob, dtype=prob.dtype)
+        test = shannon_entropy(prob, backend=backend)
+    with pytest.raises(TypeError):
+        prob = np.array([])
+        prob = backend.cast(prob, dtype=prob.dtype)
+        test = shannon_entropy(prob, backend=backend)
+    with pytest.raises(ValueError):
+        prob = np.array([1.0, -1.0])
+        prob = backend.cast(prob, dtype=prob.dtype)
+        test = shannon_entropy(prob, backend=backend)
+    with pytest.raises(ValueError):
+        prob = np.array([1.1, 0.0])
+        prob = backend.cast(prob, dtype=prob.dtype)
+        test = shannon_entropy(prob, backend=backend)
+    with pytest.raises(ValueError):
+        prob = np.array([0.5, 0.4999999])
+        prob = backend.cast(prob, dtype=prob.dtype)
+        test = shannon_entropy(prob, backend=backend)
+
+
+@pytest.mark.parametrize("base", [2, 10, np.e, 5])
+def test_shannon_entropy(backend, base):
+    prob_array = [1.0, 0.0]
+    result = shannon_entropy(prob_array, base, backend=backend)
+    backend.assert_allclose(result, 0.0)
+
+    if base == 2:
+        prob_array = np.array([0.5, 0.5])
+        prob_array = backend.cast(prob_array, dtype=prob_array.dtype)
+        result = shannon_entropy(prob_array, base, backend=backend)
+        backend.assert_allclose(result, 1.0)
 
 
 @pytest.mark.parametrize("check_hermitian", [False, True])
